Add page size parameter (#1004)
* Add page size parameter

* Update src/Altinn.Notifications.Core/Configuration/StatusFeedConfig.cs


* Update src/Altinn.Notifications.Core/Services/StatusFeedService.cs


* Add changes based on feedback

* Add changes based on feedback

* Add validator for status feed request

* Update src/Altinn.Notifications.Core/Configuration/NotificationConfig.cs


* Update src/Altinn.Notifications.Core/Models/Status/GetStatusFeedRequest.cs


* Fix seq range

* Add test class for getstatusfeedrequest

* Add changes based on feedback

* Update data type to long for tests

* Update test/Altinn.Notifications.Tests/Notifications.Core/TestingServices/StatusFeedServiceTests.cs


---------

+++ src/Altinn.Notifications/Models/Status/StatusFeedExt.cs
@@ -11,5 +11,5 @@ public record StatusFeedExt : OrderStatusExt
     /// The sequence number of the status feed 
     /// </summary>
     [JsonPropertyName("sequenceNumber")]
-    public required int SequenceNumber { get; init; }
+    public required long SequenceNumber { get; init; }
 }
